+++ inlaprog/src/inla-special-functions.c
@@ -103,24 +103,65 @@ double inla_logcdf_normal_fast(double x)
 	}
 }
 
-double inla_lgamma_fast(double x)
+double inla_lgamma_fast1(double x)
 {
-	// this is the Gergo Nemes (2007) approximation from
-	// https://en.wikipedia.org/wiki/Stirling's_approximation
+	// this is the G.Nemes (2007) approximation from https://en.wikipedia.org/wiki/Stirling's_approximation
 
 	if (round(x) == x) {
-		return (gsl_sf_lngamma(x));
+		return gsl_sf_lnfact((int)x - 1);
 	}
 
 	double val;
 	if (x < 1.0) {
-		val = gsl_sf_lngamma(x);
+		val = LGAMMAfn(x);
 	} else {
 		double lx = log(x);
 		val = 0.5 * (LOG2PI - lx) + x * (log(x + 1.0 / (12.0 * x - 0.1 / x)) - 1.0);
 	}
 	return (val);
 }
+
+double inla_gamma_fast1(double x) 
+{
+	return (exp(inla_lgamma_fast1(x)));
+}
+
+double inla_lgamma_fast2(double x)
+{
+	if (unlikely(x <= 0.0)) {
+		return lgamma(x);
+	}
+
+#define G 5
+#define N 7
+	static const double p[N] =  {
+		1.0000000001900148240,
+		76.180091729471463483,
+		-86.505320329416767652,
+		24.014098240830910490,
+		-1.2317395724501553875,
+		0.0012086509738661785061,
+		-5.3952393849531283785e-6
+	};
+	
+	double tmp = x + G + 0.5;
+	tmp -= (x + 0.5) * log(tmp);
+    
+	double ser = p[0];
+	for (int i = 1; i < N; ++i) {
+		ser += p[i] / (x+i);
+	}
+    
+#undef G
+#undef N	
+	return -tmp + log(2.5066282746310005 * ser / x);
+}
+
+double inla_gamma_fast2(double x) 
+{
+	return (exp(inla_lgamma_fast2(x)));
+}
+
 
 double inla_ipow(double x, int k)
 {